DGS-3246 вынес пару интерфейсов
