+++ src/ClientInterface.php
@@ -2,32 +2,15 @@
 
 namespace Dooglys\Api;
 
+use Dooglys\Api\Module\LoyaltyClientInterface;
+use Dooglys\Api\Module\StructureClientInterface;
+
 /**
  * Interface ClientInterface
  *
  * @package Dooglys\Api
  */
-interface ClientInterface {
-    /**
-     * API метод /v1/structure/sale-point/view
-     *
-     * @param $id
-     *
-     * @return mixed
-     *
-     */
-    public function structureSalePointView ($id);
-
-    /**
-     * API метод /v1/structure/sale-point/list
-     *
-     * @param $options
-     *
-     * @return mixed
-     *
-     */
-    public function structureSalePointList (array $options = []);
-
+interface ClientInterface extends LoyaltyClientInterface, StructureClientInterface {
     /**
      * API метод /v1/nomenclature/product/view
      *
@@ -69,75 +52,12 @@ interface ClientInterface {
     public function nomenclatureCategoryList (array $options = []);
 
     /**
-     * API метод /v1/structure/user/view
-     * @param $id
-     *
-     * @return mixed
-     */
-    public function structureUserView ($id);
-
-    /**
-     * API метод /v1/structure/user/list
-     * @param array $options
-     *
-     * @return mixed
-     */
-    public function structureUserList (array $options = []);
-
-    /**
      * API метод /v1/terminal-menu/menu/view
      * @param $id
      *
      * @return mixed
      */
     public function terminalMenuMenuView ($id);
-
-    /**
-     * API метод /v1/loyalty/transaction/buy-commit
-     * @param array $options
-     *
-     * @return mixed
-     */
-    public function loyaltyTransactionBuyCommit (array $options = []);
-
-    /**
-     * API метод /v1/loyalty/transaction/buy-new
-     * @param array $options
-     *
-     * @return mixed
-     */
-    public function loyaltyTransactionBuyNew (array $options = []);
-
-    /**
-     * API метод /v1/loyalty/transaction/buy-return
-     * @param array $options
-     *
-     * @return mixed
-     */
-    public function loyaltyTransactionBuyReturn (array $options = []);
-
-    /**
-     * API метод /v1/loyalty/card/info
-     * @param $options
-     *
-     * @return mixed
-     */
-    public function loyaltyCardInfo ($options);
-
-    /**
-     * API метод /v1/loyalty/settings/view
-     *
-     * @return mixed
-     */
-    public function loyaltySettingsView ();
-
-    /**
-     * API метод /v1/structure/tenant/settings
-     * @param array $options
-     *
-     * @return mixed
-     */
-    public function structureTenantSettings (array $options = []);
 
     /**
      * API метод /v1/terminal-menu/menu/kit
